# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Overtime Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List employees, optionally filtered by roster
      parameters:
        - name: rosterId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create a new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Updated employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /employees/{id}/report:
    get:
      operationId: getEmployeeReport
      tags: [reports]
      summary: Get overtime history for an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeReport"
        "404":
          description: Not found

  # ── Events ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Employee:
      type: object
      required: [id, rosterId, name, seniority, active]
      properties:
        id:
          type: integer
        rosterId:
          type: integer
        name:
          type: string
        seniority:
          type: integer
        roleId:
          type: ["integer", "null"]
        roleName:
          type: ["string", "null"]
        subclassId:
          type: ["integer", "null"]
        subclassName:
          type: ["string", "null"]
        active:
          type: boolean
        totalOfferedHours:
          type: number
        totalWorkedHours:
          type: number
        fairnessScore:
          type: number
          description: Weighted offered hours used for rotation ordering
        startingNormalizedHours:
          type: number
          description: Baseline hours used for fairness computation; not counted in offered/worked totals
    EmployeeInput:
      type: object
      required: [rosterId, name, seniority]
      properties:
        rosterId:
          type: integer
        name:
          type: string
          minLength: 1
        seniority:
          type: integer
          minimum: 1
        roleId:
          type: ["integer", "null"]
        subclassId:
          type: ["integer", "null"]
        active:
          type: boolean
          default: true
        startingNormalizedHours:
          type: number
          minimum: 0
    EmployeeUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        seniority:
          type: integer
          minimum: 1
        roleId:
          type: ["integer", "null"]
        subclassId:
          type: ["integer", "null"]
        active:
          type: boolean
        startingNormalizedHours:
          type: number
          minimum: 0

    # ── Events ────────────────────────────────────────────────────────────────
...
    EmployeeReport:
      type: object
      required: [employee, totalOfferedHours, totalWorkedHours, events]
      properties:
        employee:
          $ref: "#/components/schemas/Employee"
        totalOfferedHours:
          type: number
        totalWorkedHours:
          type: number
        acceptanceRate:
          type: number
        events:
          type: array
          items:
            $ref: "#/components/schemas/EmployeeEventRecord"
    EmployeeEventRecord:
      type: object
      required: [eventId, date, description, dayType, offered, worked, hoursOffered, hoursAwarded, multiplier]
      properties:
        eventId:
          type: integer
        date:
          type: string
          format: date
        description:
          type: string
        dayType:
          type: string
        offered:
          type: boolean
        worked:
          type: boolean
        hoursOffered:
          type: number
        hoursAwarded:
          type: number
        hoursOverride:
          type: ["number", "null"]
        multiplier:
          type: number
          default: 1
...