# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Overtime Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Employee {
  id: number;
  rosterId: number;
  name: string;
  seniority: number;
  /** @nullable */
  roleId?: number | null;
  /** @nullable */
  roleName?: string | null;
  /** @nullable */
  subclassId?: number | null;
  /** @nullable */
  subclassName?: string | null;
  active: boolean;
  totalOfferedHours?: number;
  totalWorkedHours?: number;
  /** Weighted offered hours used for rotation ordering */
  fairnessScore?: number;
  /** Baseline hours used for fairness computation; not counted in offered/worked totals */
  startingNormalizedHours?: number;
}