# lib/api-zod/src/generated/types/employeeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Overtime Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface EmployeeInput {
  rosterId: number;
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  seniority: number;
  /** @nullable */
  roleId?: number | null;
  /** @nullable */
  subclassId?: number | null;
  active?: boolean;
  /** @minimum 0 */
  startingNormalizedHours?: number;
}